Create electricity_bill.c
An electricity board charges the following rates for the use of electricity - for the first 200 units 80 paise per unit ; for the next 100 units 90 paise per unit ; beyond 300 units Rs 1 per unit. 
All users are charged a minimum of Rs. 100 as meter charge. If the total amount is more than Rs 400, then an additional surcharge of 15% of total amount is charged. A program to read the name of the user, number of units consumed and print out the charges.